+++ src/main/java/mdteam/ait/core/AITSounds.java
@@ -7,8 +7,9 @@ import net.minecraft.sound.SoundEvent;
 import net.minecraft.util.Identifier;
 
 public class AITSounds implements NeptuneSoundEventInit {
-    public static final SoundEvent SECRET_MUSIC = SoundEvent.of(new Identifier(AITMod.MOD_ID, "secret_music"));
-    public static final SoundEvent EVEN_MORE_SECRET_MUSIC = SoundEvent.of(new Identifier(AITMod.MOD_ID, "even_more_secret_music"));
+    public static final SoundEvent SECRET_MUSIC = SoundEvent.of(new Identifier(AITMod.MOD_ID, "music/secret_music"));
+    public static final SoundEvent EVEN_MORE_SECRET_MUSIC = SoundEvent.of(new Identifier(AITMod.MOD_ID, "music/even_more_secret_music"));
+    public static final SoundEvent DRIFTING_MUSIC = SoundEvent.of(new Identifier(AITMod.MOD_ID, "music/drifting_by_radio"));
 
     // TARDIS
     public static final SoundEvent DEMAT = SoundEvent.of(new Identifier(AITMod.MOD_ID, "tardis/demat"));