+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 - Changed how crashing works so that the player actually takes damage when they crash severity depends on annoyance
 - Fixed a crash that happens when the tardis loses power
 - Fixed a crash in server tardis manager
+- Add a music section for sounds
+- Added drifting song by radio (I'll update with the new version when he adds that)